comments, FIEO, SQL Injection, CSRF tokens

+++ delete_comment.php
@@ -8,12 +8,16 @@
     <title>Document</title>
 </head>
 <body>
-
-    <?php  
+<!--delete a comment from database-->
+    <?php 
     session_start();
+    if(!hash_equals($_SESSION['token'], $_POST['token'])){
+        die("Request forgery detected");
+    }
     $comment_id=$_POST['comment_to_delete'];
     echo $comment_id;
     require 'database.php';
+    //prepare to delete from table comments
     $stmt = $mysqli->prepare("delete from comments where comment_id=?");
     if(!$stmt){
         printf("Query Prep Failed: %s\n", $mysqli->error);
